Updated v1 schemas with v2 params, deleted v2 files
