+++ app/microservice/register.json
@@ -88,69 +88,6 @@
           "path": "/api/v1/user/obtain/all-users"
         }
       ]
-    }, {
-      "url": "/v2/user",
-      "authenticated": true,
-      "method": "GET",
-      "endpoints": [{
-          "method": "GET",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user"
-      }]
-  }, {
-      "url": "/v2/user",
-      "method": "POST",
-      "authenticated": true,
-      "endpoints": [{
-          "method": "POST",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user"
-      }]
-  }, {
-      "url": "/v2/user/:id",
-      "method": "GET",
-      "authenticated": true,
-      "endpoints": [{
-          "method": "GET",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user/:id"
-      }]
-  }, {
-      "url": "/v2/user/stories",
-      "method": "GET",
-      "authenticated": true,
-      "endpoints": [{
-          "method": "GET",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user/stories"
-      }]
-  }, {
-      "url": "/v2/user/:id",
-      "method": "DELETE",
-      "authenticated": true,
-      "endpoints": [{
-          "method": "DELETE",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user/:id"
-      }]
-  }, {
-      "url": "/v2/user/:id",
-      "method": "PATCH",
-      "authenticated": true,
-      "endpoints": [{
-          "method": "PATCH",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user/:id"
-      }]
-  }, {
-      "url": "/v2/user/obtain/all-users",
-      "authenticated": true,
-      "method": "GET",
-      "endpoints": [{
-          "method": "GET",
-          "baseUrl": "#(service.uri)",
-          "path": "/api/v2/user/obtain/all-users"
-      }]
-  }
+    }
   ]
 }
