+++ app/src/models/user.js
@@ -8,14 +8,23 @@ const User = new Schema({
     email: { type: String, required: false, trim: true },
     createdAt: { type: Date, required: false, default: Date.now },
     sector: { type: String, required: false, trim: true },
-    primaryResponsibilities: { type: Array, default: [] },
     country: { type: String, required: false, trim: true },
     state: { type: String, required: false, trim: true },
     city: { type: String, required: false, trim: true },
     howDoYouUse: { type: Array, default: [] },
+    primaryResponsibilities: { type: Array, default: [] },
     signUpForTesting: { type: Boolean, default: false },
     language: { type: String, required: false },
-    profileComplete: { type: Boolean, required: true, default: false }
+    profileComplete: { type: Boolean, required: true, default: false },
+    subsector: { type: String, required: false, trim: true },
+    jobTitle: { type: String, trim: true },
+    company: { type: String, trim: true },
+    aoiCountry: { type: String, required: true, trim: true },
+    aoiState: { type: String, required: true, trim: true },
+    aoiCity: { type: String, required: true, trim: true },
+    interests: { type: Array, default: [] },
+    signUpToNewsletter: { type: Boolean, default: false },
+    topics: { type: Array, default: [] }
 });
 
 module.exports = mongoose.model('User', User);
