+++ app/test/e2e/utils/helpers.js
@@ -10,6 +10,9 @@ const createUser = (anotherData = {}) => {
         state: `Fake state ${uuid}`,
         country: `Fake country ${uuid}`,
         city: `Fake city ${uuid}`,
+        aoiCountry: 'Fake AOI country',
+        aoiState: 'Fake AOI state',
+        aoiCity: 'Fake AOI city',
         primaryResponsibilities: ['fake responsability'],
         howDoYouUse: ['fake howDoYouUse'],
         profileComplete: true,
@@ -19,26 +22,6 @@ const createUser = (anotherData = {}) => {
     };
 };
 
-const createUserV2 = (anotherData = {}) => {
-    const uuid = mongoose.Types.ObjectId();
-
-    return {
-        firstName: `Fake firstname ${uuid}`,
-        lastName: `Fake lastname ${uuid}`,
-        email: `fake-email-${uuid}@example.com`,
-        aoiCountry: 'Fake AOI country',
-        aoiState: 'Fake AOI state',
-        aoiCity: 'Fake AOI city',
-        sector: `Fake sector ${uuid}`,
-        state: `Fake state ${uuid}`,
-        country: `Fake country ${uuid}`,
-        city: `Fake city ${uuid}`,
-        howDoYouUse: ['fake howDoYouUse'],
-        ...anotherData
-    };
-};
-
 module.exports = {
-    createUser,
-    createUserV2
+    createUser
 };
